% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline_model.R
\name{compute_expected_cases}
\alias{compute_expected_cases}
\title{Compute expected cases}
\usage{
compute_expected_cases(beta, Q, lnPmat, Tobs)
}
\arguments{
\item{beta}{parameter vector of num_params}

\item{Q}{spline basis matrix, of size Tmod x num_params}

\item{lnPmat}{matrix size Tobs x Tobs, log of make_likelihood_matrix}

\item{Tobs}{maximum observed time point}
}
\value{
A Tobs-length vector that models expected cases
}
\description{
This function computes expected cases given incidence curve parameters and a delay distribution.
}
