% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covid_new_york_city}
\alias{covid_new_york_city}
\title{New York City data from the COVID-19 pandemic.}
\format{
A data frame with 615 entries and 5 columns.
\describe{
  \item{date}{record date}
  \item{borough}{record borough: Brooklyn, Bronx, Manhattan, Queens, and Staten Island}
  \item{case}{number of recorded cases}
  \item{hospitalization}{number of new hospital admissions}
  \item{death}{number of recorded deaths}
}
}
\source{
New York City Department of Health \url{https://raw.githubusercontent.com/nychealth/coronavirus-data/master/boro/boroughs-case-hosp-death.csv}.
}
\usage{
covid_new_york_city
}
\description{
Daily case, hospitalization, and death proportions by borough through 2020-06-30.
}
\keyword{datasets}
