% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline_model.R
\name{diff_trans}
\alias{diff_trans}
\title{Transpose of the 1st difference operator}
\usage{
diff_trans(a)
}
\arguments{
\item{a}{A vector of inputs}
}
\value{
The transpose of the first difference operator
}
\description{
This function computes a transpose of the 1st difference operator.
}
