% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_processing.R
\name{front_zero_pad}
\alias{front_zero_pad}
\title{Pad reported data with zeros in front}
\usage{
front_zero_pad(reported, size)
}
\arguments{
\item{reported}{An integer vector of reported cases}

\item{size}{An integer size of zero-padding}
}
\value{
An integer vector of cases with size 0's in front
}
\description{
Add zeros in front of reported data avoid infections from before first reported date all being placed on first reported date.
}
