% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline_model.R
\name{init_params}
\alias{init_params}
\title{Initialize spline parameters (beta)}
\usage{
init_params(num_params)
}
\arguments{
\item{num_params}{Integer size of desired parameter vector}
}
\value{
vector of size num_params
}
\description{
Initialize spline parameters (beta) using a standard Gaussian distribution.
}
