% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline_model.R
\name{make_likelihood_matrix}
\alias{make_likelihood_matrix}
\title{Make delay likelihood matrix}
\usage{
make_likelihood_matrix(delay_dist)
}
\arguments{
\item{delay_dist}{A positive vector that sums to one, which describes the delay distribution.}
}
\value{
A matrix of size n x n
}
\description{
This function creates a matrix such that P[t, s] = P(C = t | I = s) = theta_{t-s} for s <= t and 0 otherwise.
}
