% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline_model.R
\name{make_spline_basis}
\alias{make_spline_basis}
\title{Create spline basis matrix}
\usage{
make_spline_basis(dof, tgrid)
}
\arguments{
\item{dof}{An integer degrees of freedom.}

\item{tgrid}{A grid of time values.}
}
\value{
A matrix of cubic spline basis values with `length(tgrid)` x `dof` entries.
}
\description{
This function creates basis matrix for spline model using cubic splines.
}
