% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline_model.R
\name{marg_loglike_poisson}
\alias{marg_loglike_poisson}
\title{Marginal log likelihood 
This function computes the marginal probability of Pr(reported | beta).  Note that
lnPmat must be zero padded enough (or censored) to match the length of 
reported cases vector.}
\usage{
marg_loglike_poisson(beta, reported, Q, lnPmat)
}
\arguments{
\item{beta}{spline parameter vector length num_params}

\item{reported}{An integer vector of reported cases.}

\item{Q}{spline basis matrix Tmod x num_params}

\item{lnPmat}{matrix size Tobs x Tobs, log of make_likelihood_matrix}
}
\value{
A scalar log likelihood value.
}
\description{
Marginal log likelihood 
This function computes the marginal probability of Pr(reported | beta).  Note that
lnPmat must be zero padded enough (or censored) to match the length of 
reported cases vector.
}
