% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline_model.R
\name{marg_loglike_poisson_fisher}
\alias{marg_loglike_poisson_fisher}
\title{Marginal log likelihood Fisher information matrix}
\usage{
marg_loglike_poisson_fisher(beta, reported, Q, lnPmat)
}
\arguments{
\item{beta}{A spline parameter vector length num_params.}

\item{reported}{An integer vector of reported cases.}

\item{Q}{A spline basis matrix Tmod x num_params.}

\item{lnPmat}{A matrix size Tobs x Tobs, log of make_likelihood_matrix.}
}
\value{
A numeric vector, gradient of log likelihood value with respect to beta.
}
\description{
This function computes the Fisher information matrix log likelihood term with
respect to beta.
}
