% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline_model.R
\name{marg_loglike_poisson_grad}
\alias{marg_loglike_poisson_grad}
\title{Marginal log likelihood gradient}
\usage{
marg_loglike_poisson_grad(beta, reported, Q, lnPmat)
}
\arguments{
\item{beta}{spline parameter vector length num_params}

\item{reported}{An integer vector of reported cases.}

\item{Q}{spline basis matrix Tmod x num_params}

\item{lnPmat}{matrix size Tobs x Tobs, log of make_likelihood_matrix}
}
\value{
A numeric vector, gradient of log likelihood value with respect to beta.
}
\description{
This function computes the gradient of the log likelihood term with respect to beta.
}
