% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{plot.incidence_spline_model}
\alias{plot.incidence_spline_model}
\title{Plot model from fit_incidence}
\usage{
\method{plot}{incidence_spline_model}(x, ...)
}
\arguments{
\item{x}{An "incidence_spline_model" output from \code{\link{fit_incidence}}.}

\item{...}{Other parameters that can be included: \itemize{
\item `times`: an optional vector of time indices.
\item `plot_Chat`: a logical for whether Chat should be plotted. 
\item `plot_reported`: a logical for whether reported cases should be plotted.
\item `plot_CI`: a logical for whether CI should be plotted.
}}
}
\description{
Plot time, reported cases, incidence curve with credible interval, and implied case curve.
}
\examples{
\donttest{
indiana_model <- fit_incidence(
                  reported = spanish_flu$Indiana, 
                  delay_dist = spanish_flu_delay_dist$proportion)
 plot(indiana_model, times = spanish_flu$Date)
}
}
