% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline_model.R
\name{poisson_objective_grad}
\alias{poisson_objective_grad}
\title{Poisson objective function gradient}
\usage{
poisson_objective_grad(beta, lam, reported, Q, lnPmat, regularization_order)
}
\arguments{
\item{beta}{spline parameter vector length num_params}

\item{lam}{positive scalar regularization strength}

\item{reported}{An integer vector of reported cases.}

\item{Q}{spline basis matrix Tmod x num_params}

\item{lnPmat}{matrix size Tobs x Tobs, log of make_likelihood_matrix}

\item{regularization_order}{An integer (typically 0, 1, 2), indicating differencing order for L2 
regularization of spline parameters. Default is 2 for second derivative penalty.}
}
\value{
scalar objective function value
}
\description{
This function computes the Poisson objective function (including regularizer) gradient.
}
