% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline_model.R
\name{poisson_objective_post_cov_approx}
\alias{poisson_objective_post_cov_approx}
\title{Compute Fisher information matrix for Poisson objective}
\usage{
poisson_objective_post_cov_approx(
  beta,
  lam,
  reported,
  Q,
  lnPmat,
  regularization_order
)
}
\arguments{
\item{beta}{A vector of spline parameters.}

\item{lam}{A regularization penalty parameter.}

\item{reported}{A vector of reported values.}

\item{Q}{A spline basis matrix.}

\item{lnPmat}{A matrix size Tobs x Tobs, log of make_likelihood_matrix.}

\item{regularization_order}{An integer that specifies the regularization order.}
}
\value{
Fisher information matrix of a regularized Poisson objective function.
}
\description{
This function computes the Fisher information matrix for a regularized Poisson objective function.
}
