% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spline_model.R
\name{regfun_grad}
\alias{regfun_grad}
\title{Beta regularization function gradient}
\usage{
regfun_grad(beta, regularization_order = 2)
}
\arguments{
\item{beta}{spline parameter vector length num_params}

\item{regularization_order}{An integer (typically 0, 1, 2), indicating differencing order for L2 
regularization of spline parameters. Default is 2 for second derivative penalty.}
}
\value{
scalar regularization value
}
\description{
This function computes regularization penalty term gradient based on the betas and difference order.
}
