% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spanish_flu}
\alias{spanish_flu}
\title{Daily flu mortality from 1918 flu pandemic.}
\format{
A data frame with 122 entries for 3 locations
\describe{
  \item{Date}{date}
  \item{Indiana}{daily deaths for all of Indiana}
  \item{Kansas}{daily deaths for all of Kansas}
  \item{Philadelphia}{daily deaths for Philadelphia}
}
}
\source{
Rogers SL (1920). Special Tables of Mortality from Influenza and Pneumonia, in Indiana, Kansas, and Philadelphia, PA (U.S. Dept Commerce, Washington, DC).
}
\usage{
spanish_flu
}
\description{
Daily mortality data from 1918-09-01 through 1918-12-31 in Indiana, Kansas, and Philadelphia
}
\keyword{datasets}
