% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{spanish_flu_delay_dist}
\alias{spanish_flu_delay_dist}
\title{Delay distribution from 1918 flu pandemic.}
\format{
A data frame with 31 entries and 3 columns.
\describe{
  \item{days}{number of days since infection}
  \item{proportion}{proportion of deaths that happen on that day}
}
}
\source{
Goldstein E, et al. (2009). Reconstructing influenza incidence by deconvolution of daily mortality time series (PNAS). \url{https://www.pnas.org/content/pnas/106/51/21825.full.pdf}
}
\usage{
spanish_flu_delay_dist
}
\description{
Daily death proportions.
}
\keyword{datasets}
