% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_fitting.R
\name{train_val_split}
\alias{train_val_split}
\title{Split reported case data}
\usage{
train_val_split(reported, frac_train = 0.75)
}
\arguments{
\item{reported}{An integer vector of reported cases.}

\item{frac_train}{A numeric between 0 and 1 for fraction of data used to train lambda validation.}
}
\value{
A list(reported_train, reported_val) where the elements reported_train
        and reported_val are both length, Tobs, and 'frac_train' of
        the counts fall in reported_train, the rest in reported_val.
}
\description{
Split reported case integer time series into train and validate time series through thinning.
}
